+++ src/Mod/Draft/CMakeLists.txt
@@ -220,6 +220,9 @@ SET(Draft_GUI_tools
     draftguitools/gui_snaps.py
     draftguitools/gui_snapper.py
     draftguitools/gui_trackers.py
+    draftguitools/gui_edit_draft_objects.py
+    draftguitools/gui_edit_arch_objects.py
+    draftguitools/gui_edit_part_objects.py
     draftguitools/gui_edit.py
     draftguitools/gui_lineops.py
     draftguitools/gui_togglemodes.py
